requirement.txt
asgiref                3.5.0    
certifi                2021.10.8
cffi                   1.15.0   
charset-normalizer     2.0.12   
cryptography           36.0.1   
defusedxml             0.7.1
Deprecated             1.2.13
Django                 3.2.11
django-annoying        0.10.6
django-cors-headers    3.11.0
django-environ         0.8.1
django-oauth-toolkit   1.7.0
django-rest-framework  0.1.0
djangorestframework    3.13.1
drf-social-oauth2      1.2.1
ecdsa                  0.17.0
idna                   3.3
jwcrypto               1.0
mysqlclient            1.4.6
oauthlib               3.2.0
Pillow                 9.0.1
pip                    22.0.3
pyasn1                 0.4.8
pycparser              2.21
PyJWT                  2.3.0
python-decouple        3.6
python-jose            3.3.0
python3-openid         3.2.0
pytz                   2021.3
requests               2.27.1
requests-oauthlib      1.3.1
rsa                    4.8
setuptools             60.2.0
six                    1.16.0
social-auth-app-django 5.0.0
social-auth-core       4.2.0
sqlparse               0.4.2
typing_extensions      4.0.1
urllib3                1.26.8
wheel                  0.37.1
wrapt                  1.13.3